@@ -1,7 +1,7 @@
 cmake_minimum_required(VERSION 3.5)
 project(raft-kv)
 
-set(CMAKE_CXX_STANDARD 11)
+set(CMAKE_CXX_STANDARD 14)
 
 add_definitions(-Wall)
 add_definitions(-DDEBUG)
