@@ -4,9 +4,13 @@
 
 namespace kv {
 
+// 对于预写日志的管理
 class RaftLog {
  public:
-  explicit RaftLog(StoragePtr storage, uint64_t max_next_ents_size);
+  explicit RaftLog(
+    StoragePtr storage,  
+    uint64_t max_next_ents_size，
+    );
 
   ~RaftLog();
 
@@ -69,6 +73,8 @@ class RaftLog {
   // later term is more up-to-date. If the logs end with the same term, then
   // whichever log has the larger lastIndex is more up-to-date. If the logs are
   // the same, the given log is up-to-date.
+
+  // 如果候选人的最后一条日志的任期大于当前节点的任期 || 候选人的最后一条日志的索引大于等于当前节点的最后一条日志的索引，则说明候选人的日志是最新的
   bool is_up_to_date(uint64_t lasti, uint64_t term) const {
     uint64_t lt = last_term();
     return term > lt || (term == lt && lasti >= last_index());
@@ -98,7 +104,7 @@ class RaftLog {
     if (index > last_index()) {
       return Status::ok();
     }
-    return slice(index, last_index() + 1, max_size, entries);
+    return slice(index, last_index() + 1, max_size, entries); // 返回从 index 到 last_index() + 1 的 entries， 如果日志已经被压缩则返回!ok
   }
 
   void commit_to(uint64_t to_commit);
@@ -118,24 +124,28 @@ class RaftLog {
   void all_entries(std::vector<proto::EntryPtr>& entries);
 
  public:
-  // storage contains all stable entries since the last snapshot.
-  StoragePtr storage_;
+   // storage contains all stable entries since the last snapshot.
+   // 已经持久化的预写日志模块，但仅仅是query预写日志，因为预写日志的变化不应该由算法层来实现，因此仅仅是提供了一个 storage 的指针
+   StoragePtr storage_;
 
-  // unstable contains all unstable entries and snapshot.
-  // they will be saved into storage.
-  UnstablePtr unstable_;
+   // unstable contains all unstable entries and snapshot.
+   // they will be saved into storage.
+   // 还未持久化在 storage 中的预写日志位置，该位置之后的预写日志还在内存中
+   UnstablePtr unstable_;  
 
-  // committed is the highest log position that is known to be in
-  // stable storage on a quorum of nodes.
-  uint64_t committed_;
-  // applied is the highest log position that the application has
-  // been instructed to apply to its state machine.
-  // Invariant: applied <= committed
-  uint64_t applied_;
+   // committed is the highest log position that is known to be in
+   // stable storage on a quorum of nodes.
+   // 记录了已经提交的预写日志的位置 commitIndex
+   uint64_t committed_;
+   // applied is the highest log position that the application has
+   // been instructed to apply to its state machine.
+   // Invariant: applied <= committed
+   // 记录了已经应用在状态机上的预写日志的位置 applyIndex，applied <= committed是因为一条日志首先要提交成功才能够应用在状态机上
+   uint64_t applied_;
 
-  // max_next_ents_size is the maximum number aggregate byte size of the messages
-  // returned from calls to nextEnts.
-  uint64_t max_next_ents_size_;
+   // max_next_ents_size is the maximum number aggregate byte size of the messages
+   // returned from calls to nextEnts.
+   uint64_t max_next_ents_size_;
 };
 typedef std::shared_ptr<RaftLog> RaftLogPtr;
 
