@@ -7,6 +7,8 @@ namespace kv {
 // Note that unstable.offset may be less than the highest log
 // position in storage; this means that the next write to storage
 // might need to truncate the log before persisting unstable.entries.
+
+// unstable存储的是还没有被持久化到 storage 的日志及快照，offset 代表了 entries 中第一条日志的 index
 class Unstable {
  public:
   explicit Unstable(uint64_t offset)
@@ -36,12 +38,11 @@ class Unstable {
 
   void slice(uint64_t low, uint64_t high, std::vector<proto::EntryPtr>& entries);
  public:
-  // the incoming unstable snapshot, if any.
+  // 非 0 的时候表示，已经有传入的快照
   proto::SnapshotPtr snapshot_;
   // all entries that have not yet been written to storage.
-
-  std::vector<proto::EntryPtr> entries_;
-  uint64_t offset_;
+  std::vector<proto::EntryPtr> entries_;  // 还未被持久化进storage中的预写日志
+  uint64_t offset_;  // 标识了entries_中的第一个entry的index
 };
 typedef std::shared_ptr<Unstable> UnstablePtr;
 
