@@ -6,7 +6,7 @@
 
 namespace kv {
 
-// see redis keys command
+// see redis keys command  返回的key支持正则匹配
 int string_match_len(const char* pattern, int patternLen,
                      const char* string, int stringLen, int nocase) {
   while (patternLen && stringLen) {
@@ -127,11 +127,13 @@ int string_match_len(const char* pattern, int patternLen,
   return 0;
 }
 
+// 初始化
 RedisStore::RedisStore(RaftNode* server, std::vector<uint8_t> snap, uint16_t port)
     : server_(server),
       acceptor_(io_service_),
       next_request_id_(0) {
 
+  // 初始化如果传入快照，则根据快照把 kv 重置
   if (!snap.empty()) {
     std::unordered_map<std::string, std::string> kv;
     msgpack::object_handle oh = msgpack::unpack((const char*) snap.data(), snap.size());
@@ -144,20 +146,23 @@ RedisStore::RedisStore(RaftNode* server, std::vector<uint8_t> snap, uint16_t por
   }
 
   auto address = boost::asio::ip::address::from_string("0.0.0.0");
-  auto endpoint = boost::asio::ip::tcp::endpoint(address, port);
+  auto endpoint = boost::asio::ip::tcp::endpoint(address, port);  // address:port
 
+  // 绑定 socket
   acceptor_.open(endpoint.protocol());
   acceptor_.set_option(boost::asio::ip::tcp::acceptor::reuse_address(1));
   acceptor_.bind(endpoint);
   acceptor_.listen();
 }
 
+// 析构函数
 RedisStore::~RedisStore() {
   if (worker_.joinable()) {
     worker_.join();
   }
 }
 
+// 启动 redisStore，promise 返回启动 RedisStore 的线程 id
 void RedisStore::start(std::promise<pthread_t>& promise) {
   start_accept();
 
@@ -170,6 +175,7 @@ void RedisStore::start(std::promise<pthread_t>& promise) {
 void RedisStore::start_accept() {
   RedisSessionPtr session(new RedisSession(this, io_service_));
 
+  // 通过 RedisSession 接受请求
   acceptor_.async_accept(session->socket_, [this, session](const boost::system::error_code& error) {
     if (error) {
       LOG_DEBUG("accept error %s", error.message().c_str());
@@ -180,28 +186,33 @@ void RedisStore::start_accept() {
   });
 }
 
+// set 命令的具体操作
 void RedisStore::set(std::string key, std::string value, const StatusCallback& callback) {
-  uint32_t commit_id = next_request_id_++;
+  uint32_t commit_id = next_request_id_++;  //  返回当前请求 id 并且增加请求的index
 
-  RaftCommit commit;
-  commit.node_id = static_cast<uint32_t>(server_->node_id());
-  commit.commit_id = commit_id;
-  commit.redis_data.type = RedisCommitData::kCommitSet;
-  commit.redis_data.strs.push_back(std::move(key));
-  commit.redis_data.strs.push_back(std::move(value));
+  RaftCommit commit;  // 提交给 raft 的结构体
+  commit.node_id = static_cast<uint32_t>(server_->node_id());  // 提交 set 请求的节点 id
+  commit.commit_id = commit_id;  // 请求id
+  commit.redis_data.type = RedisCommitData::kCommitSet;  // commit 的命令设置为 set
+  commit.redis_data.strs.push_back(std::move(key));  // push key
+  commit.redis_data.strs.push_back(std::move(value));  // push data
 
+  // 把 commit 压缩进缓冲区中
   msgpack::sbuffer sbuf;
   msgpack::pack(sbuf, commit);
+  // 把缓冲区中的数据存进 vector 中
   std::shared_ptr<std::vector<uint8_t>> data(new std::vector<uint8_t>(sbuf.data(), sbuf.data() + sbuf.size()));
 
-  pending_requests_[commit_id] = callback;
+  pending_requests_[commit_id] = callback;  // 注册对应请求的回调函数，set 命令完成后需要做的事情
 
+  // RaftNode 推进 raft propose 一笔日志，raft 层处理后返回处理结果 status，根据 status 执行回调函数
   server_->propose(std::move(data), [this, commit_id](const Status& status) {
     io_service_.post([this, status, commit_id]() {
-      if (status.is_ok()) {
+      if (status.is_ok()) {  // ok 则直接返回
         return;
       }
 
+      // 不 ok 则找到对应的请求 id，执行RedisSession 传入的 callback 函数，并且擦除 RedisStore 中的请求记录
       auto it = pending_requests_.find(commit_id);
       if (it != pending_requests_.end()) {
         it->second(status);
@@ -211,13 +222,14 @@ void RedisStore::set(std::string key, std::string value, const StatusCallback& c
   });
 }
 
+// 与上面set 命令基本一致，对请求进行记录等等，然后调用 RadtNode 的 propose 将 entry 推到 raft_层
 void RedisStore::del(std::vector<std::string> keys, const StatusCallback& callback) {
   uint32_t commit_id = next_request_id_++;
 
   RaftCommit commit;
   commit.node_id = static_cast<uint32_t>(server_->node_id());
   commit.commit_id = commit_id;
-  commit.redis_data.type = RedisCommitData::kCommitDel;
+  commit.redis_data.type = RedisCommitData::kCommitDel;  // 记录命令类型是 del，与 set 不同
   commit.redis_data.strs = std::move(keys);
   msgpack::sbuffer sbuf;
   msgpack::pack(sbuf, commit);
@@ -242,10 +254,11 @@ void RedisStore::get_snapshot(const GetSnapshotCallback& callback) {
     msgpack::sbuffer sbuf;
     msgpack::pack(sbuf, this->key_values_);
     SnapshotDataPtr data(new std::vector<uint8_t>(sbuf.data(), sbuf.data() + sbuf.size()));
-    callback(data);
+    callback(data);  // redis 获取到快照的数据后，根据 raft_node 传入的函数处理 data
   });
 }
 
+// 将 snapshot 中的 kv 和 redisStore 中的 key_values_ 交换，从而实现从快照恢复 redisStore 中的 key_values_
 void RedisStore::recover_from_snapshot(SnapshotDataPtr snap, const StatusCallback& callback) {
   io_service_.post([this, snap, callback] {
     std::unordered_map<std::string, std::string> kv;
@@ -256,19 +269,21 @@ void RedisStore::recover_from_snapshot(SnapshotDataPtr snap, const StatusCallbac
       callback(Status::io_error("invalid snapshot"));
       return;
     }
-    std::swap(kv, key_values_);
-    callback(Status::ok());
+    std::swap(kv, key_values_);  // 从快照中恢复
+    callback(Status::ok());  // 回调函数 ok
   });
 }
 
+// 回复所有的 keys，支持正则匹配
 void RedisStore::keys(const char* pattern, int len, std::vector<std::string>& keys) {
   for (auto it = key_values_.begin(); it != key_values_.end(); ++it) {
-    if (string_match_len(pattern, len, it->first.c_str(), it->first.size(), 0)) {
+    if (string_match_len(pattern, len, it->first.c_str(), it->first.size(), 0)) {  // 遍历k，符合模式匹配的加入到 keys 中
       keys.push_back(it->first);
     }
   }
 }
 
+//  entry中具体对 redis 的操作
 void RedisStore::read_commit(proto::EntryPtr entry) {
   auto cb = [this, entry] {
     RaftCommit commit;
@@ -284,13 +299,13 @@ void RedisStore::read_commit(proto::EntryPtr entry) {
     RedisCommitData& data = commit.redis_data;
 
     switch (data.type) {
-      case RedisCommitData::kCommitSet: {
-        assert(data.strs.size() == 2);
-        this->key_values_[std::move(data.strs[0])] = std::move(data.strs[1]);
+      case RedisCommitData::kCommitSet: {  // redis 的 set 操作
+        assert(data.strs.size() == 2);  // kv 对
+        this->key_values_[std::move(data.strs[0])] = std::move(data.strs[1]); // 设置kv
         break;
       }
-      case RedisCommitData::kCommitDel: {
-        for (const std::string& key : data.strs) {
+      case RedisCommitData::kCommitDel: {  // redis 的 delete 操作
+        for (const std::string& key : data.strs) {  // 删除kv
           this->key_values_.erase(key);
         }
         break;
@@ -300,16 +315,17 @@ void RedisStore::read_commit(proto::EntryPtr entry) {
       }
     }
 
-    if (commit.node_id == server_->node_id()) {
-      auto it = pending_requests_.find(commit.commit_id);
+    if (commit.node_id == server_->node_id()) {  // 提交节点的 id 为自身
+      auto it = pending_requests_.find(commit.commit_id);  // redis 自身也会维护一个 map，存放请求与对应的处理函数
       if (it != pending_requests_.end()) {
         it->second(Status::ok());
-        pending_requests_.erase(it);
+        pending_requests_.erase(it);  // 去除该请求，表明已经处理
       }
     }
   };
 
-  io_service_.post(std::move(cb));
+  // post方法是一个异步方法，调用后会马上返回，它会向io_service请求分发任务，io_serivce会选择一个线程执行。
+  io_service_.post(std::move(cb));  // 调用通信模块
 }
 
 }
